RF: Move mask2rois from datahandler to roitools
