+++ fissa/datahandler.py
@@ -58,46 +58,6 @@ def getmean(data):
     return data.mean(axis=0)
 
 
-def rois2masks(rois, data):
-    """Take the object `rois` and returns it as a list of binary masks.
-
-    Parameters
-    ----------
-    rois : string or list of array_like
-        Either a string with imagej roi zip location, list of arrays encoding
-        polygons, or list of binary arrays representing masks
-    data : array
-        Data array as made by image2array. Must be shaped `(frames, y, x)`.
-
-    Returns
-    -------
-    list
-        List of binary arrays (i.e. masks)
-
-    """
-    # get the image shape
-    shape = data.shape[1:]
-
-    # if it's a list of strings
-    if isinstance(rois, basestring):
-        rois = roitools.readrois(rois)
-
-    if not isinstance(rois, collections.Sequence):
-        raise TypeError(
-            'Wrong ROIs input format: expected a list or sequence, but got'
-            ' a {}'.format(rois.__class__)
-        )
-
-    # if it's a something by 2 array (or vice versa), assume polygons
-    if np.shape(rois[0])[1] == 2 or np.shape(rois[0])[0] == 2:
-        return roitools.getmasks(rois, shape)
-    # if it's a list of bigger arrays, assume masks
-    elif np.shape(rois[0]) == shape:
-        return rois
-
-    raise ValueError('Wrong ROIs input format: unfamiliar shape.')
-
-
 def extracttraces(data, masks):
     """Extracts a temporal trace for each spatial mask.
 
@@ -128,3 +88,7 @@ def extracttraces(data, masks):
         out[i, :] = data[:, masks[i]].mean(axis=1)
 
     return out
+
+
+# backward compatability
+rois2masks = roitools.rois2masks
